QC inspection machine modification
This machine had a various issues when operating. the machine use for inspect the fabric quality. It has operator side and fabric end side. the operators inquire fabrics are not processing properly and the machine speed is too low and also the light system of the machine need to be modify. 
Then I considered all of  the requirements and propose to modify the machine. In that modification I have add some features to the machine. there are,
1) Motorized one roller, 
2) Glued a sticker velvet with a rough surface
3) Modified the machine